加行事曆可視層級，與某人修改的patientinfo與product

+++ prjRehabilitation/Models/TScheduleDetail.cs
@@ -5,7 +5,7 @@ namespace prjRehabilitation.Models
 {
     public partial class TScheduleDetail
     {
-        public byte FId { get; set; }
+        public int FId { get; set; }
         public bool? FDeleteBool { get; set; }
         public int FGroupActivityId { get; set; }
         public string FStartTime { get; set; } = null!;
